refactor: fallback as dataclass
